requirements.txt
gradio
gradio_imageslider
torch
torchvision
scikit-image
pillow
numpy
typing
gitpython
huggingface_hub